add DataHelper wrapper for viz
