+++ cosmic_rAI/data_viz.py
@@ -1,5 +1,6 @@
 import matplotlib.pyplot as plt
 import seaborn as sns
+import numpy as np
 from .data_prep import get_charges_with_metadata
